setup/requirements.txt
click==8.1.3
docopt==0.6.2
Flask==2.2.3
importlib-metadata==6.1.0
itsdangerous==2.1.2
MarkupSafe==2.1.2
proton==0.9.1
python-qpid-proton==0.38.0
six==1.16.0
stomp.py==4.1.21
websocket-client==0.47.0
Werkzeug==2.2.3
zipp==3.15.0
